requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Emby API
requests>=2.32.0
httpx>=0.28.0
aiohttp>=3.9.0

# Video processing
opencv-python>=4.12.0
scenedetect>=0.6.6
numpy>=2.2.0

# AI providers
lmstudio>=1.4.1
ollama>=0.5.1

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Scheduling
apscheduler>=3.11.0

# CLI
click>=8.1.0
rich>=13.0.0

# Logging
structlog>=24.0.0

# Utilities
psutil>=7.0.0
aiofiles>=23.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
aiosqlite>=0.21.0
